fix(clippy): auto-fix batch for katgpt-{core,dec,types,sleep,transformer}
cargo clippy --fix across non-katgpt-rs crates. Covers needless_range_loop,
manual_contains, unnecessary_cast, iter_cloned_collect, manual_swap,
manual_is_multiple_of, manual_div_ceil, collapsible_if, and related
mechanical lints. katgpt-rs crate deferred (blocked by a clippy --fix
misfire on cs_kv_probe/lasso.rs match-bool pattern).

+++ crates/katgpt-core/src/linalg/tucker.rs
@@ -378,7 +378,7 @@ impl TuckerResultScratch {
             factor_rows: cfg.shape,
             factor_cols: cfg.ranks,
             core_shape: cfg.ranks,
-            n_modes: cfg.n_modes as u8,
+            n_modes: cfg.n_modes,
         }
     }
